Teraform/cloud-functions/DataAnalytics/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-cloud-bigquery